trading/prediction-markets/requirements.txt
# Core dependencies
py-clob-client>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
web3>=6.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Utilities
pydantic>=2.0.0
httpx>=0.25.0

# Optional: for monitoring/notifications
# discord.py>=2.3.0
# python-telegram-bot>=20.0
